created footer component
