+++ src/components/atoms/index.ts
@@ -1,17 +1,19 @@
 import BookmarkIcon from './bookmark-icon/bookmark-icon.tsx';
 import GoodListItem from './gool-list-item/gool-list-item.tsx';
-import LogoIcon from './logo-icon/logo-icon.tsx';
+import HeaderLogoIcon from './header-logo-icon/header-logo-icon.tsx';
 import PremiumLabel from './premium-label/premium-label.tsx';
 import SmallInput from './small-input/small-input.tsx';
 import SortSelectorItem from './sort-selector-item/sort-selector-item.tsx';
 import TabButton from './tab-button/tab-button.tsx';
+import LogoIcon from './logo-icon/logo-icon.tsx';
 
 export {
   BookmarkIcon,
   GoodListItem,
-  LogoIcon,
+  HeaderLogoIcon,
   PremiumLabel,
   SmallInput,
   SortSelectorItem,
   TabButton,
+  LogoIcon,
 };
