+++ src/components/organisms/index.ts
@@ -1,5 +1,7 @@
 import Header from './header/header.tsx';
+import Footer from './footer/footer.tsx';
 
 export {
   Header,
+  Footer,
 };
