+++ src/components/pages/favorites-empty/favorites-empty.tsx
@@ -1,4 +1,4 @@
-import { Header } from '../../organisms';
+import { Footer, Header } from '../../organisms';
 import { ReactElement } from 'react';
 
 function FavoritesEmptyScreen(): ReactElement {
@@ -18,17 +18,7 @@ function FavoritesEmptyScreen(): ReactElement {
           </section>
         </div>
       </main>
-      <footer className="footer">
-        <a className="footer__logo-link" href="main.html">
-          <img
-            className="footer__logo"
-            src="img/logo.svg"
-            alt="6 cities logo"
-            width={64}
-            height={33}
-          />
-        </a>
-      </footer>
+      <Footer/>
     </div>
 
   );
